extract api_key_permissions to a different file
